For-Anson/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pywebview>=6.0",
    "requests>=2.32.4",
    "werkzeug>=3.1.3",
]
